Moving login into auth component folder
