+++ src/client/components/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./change-password";
+export * from "./login";
