+++ src/client/components/index.ts
@@ -1,6 +1,6 @@
 export * from "./show-class";
 export * from "./home";
-export * from "./login";
+export * from "./auth";
 export * from "./navbar";
 export * from "./rider";
 export * from "./admin";
