+++ src/client/pages/login/index.ts
@@ -1,5 +1,5 @@
 import Vue from "vue";
-import { LoginBox } from "../../components/login";
+import { LoginBox } from "../../components/auth/login";
 
 export const LoginPage = Vue.extend({
     template: require("./login.html"),
